+++ src/panels/lovelace/entity-rows/types.ts
@@ -1,0 +1,12 @@
+import { HomeAssistant } from "../../../types";
+
+export interface EntityConfig {
+  entity: string;
+  name: string;
+  icon: string;
+}
+
+export interface EntityRow {
+  hass: HomeAssistant;
+  config: EntityConfig;
+}
