+++ src/panels/lovelace/types.ts
@@ -1,5 +1,3 @@
-import { HomeAssistant } from "../../types";
-
 export interface LovelaceConfig {
   type: string;
 }
@@ -8,15 +6,4 @@ export interface LovelaceCard extends HTMLElement {
   hass?: HomeAssistant;
   getCardSize(): number;
   setConfig(config: LovelaceConfig): void;
-}
-
-export interface EntityConfig {
-  entity: string;
-  name: string;
-  icon: string;
-}
-
-export interface EntityRow {
-  hass: HomeAssistant;
-  config: EntityConfig;
 }
